backend/requirements.txt
flask>=2.3.0
flask-cors==4.0.0
python-dotenv>=1.0.0
pypdf>=3.15.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
rank-bm25>=0.2.2
chromadb>=0.4.0
openai>=1.0.0
werkzeug>=2.3.0
torch>=2.0.0
PyJWT>=2.10.0
flask-limiter>=4.0.0
python-magic>=0.4.27
pytest>=7.0.0
pytest-cov>=4.0.0
freezegun>=1.2.0